backend-flask/requirements.txt
flask
flask-cors

opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-http
opentelemetry-instrumentation
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-requests

aws-xray-sdk

# watchtower

rollbar
blinker

PyJWT[crypto]==2.8.0
requests==2.31.0

psycopg[binary]
psycopg[pool]

